# deegree-ogcapi-features/src/main/java/org/deegree/services/oaf/resource/Feature.java
package org.deegree.services.oaf.resource;

import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.enums.ParameterStyle;
import io.swagger.v3.oas.annotations.tags.Tag;
import org.deegree.services.oaf.RequestFormat;
import org.deegree.services.oaf.exceptions.InternalQueryException;
import org.deegree.services.oaf.exceptions.InvalidParameterValue;
import org.deegree.services.oaf.exceptions.UnknownCollectionId;
import org.deegree.services.oaf.exceptions.UnknownDatasetId;
import org.deegree.services.oaf.feature.FeatureResponse;
import org.deegree.services.oaf.feature.FeatureResponseUtil;
import org.deegree.services.oaf.link.Link;
import org.deegree.services.oaf.link.LinkBuilder;
import org.deegree.services.oaf.workspace.DataAccess;
import org.deegree.services.oaf.workspace.DataAccessFactory;

import javax.ws.rs.GET;
import javax.ws.rs.Path;
import javax.ws.rs.PathParam;
import javax.ws.rs.Produces;
import javax.ws.rs.QueryParam;
import javax.ws.rs.core.Context;
import javax.ws.rs.core.Response;
import javax.ws.rs.core.UriInfo;
import java.util.Date;

import static javax.ws.rs.core.MediaType.TEXT_HTML;
import static org.deegree.services.oaf.OgcApiFeaturesConstants.HEADER_Link;
import static org.deegree.services.oaf.OgcApiFeaturesConstants.HEADER_NUMBER_MATCHED;
import static org.deegree.services.oaf.OgcApiFeaturesConstants.HEADER_NUMBER_RETURNED;
import static org.deegree.services.oaf.OgcApiFeaturesConstants.HEADER_TIMESTAMP;
import static org.deegree.services.oaf.OgcApiFeaturesMediaType.APPLICATION_GEOJSON;
import static org.deegree.services.oaf.OgcApiFeaturesMediaType.APPLICATION_GML;
import static org.deegree.services.oaf.RequestFormat.HTML;
import static org.deegree.services.oaf.RequestFormat.JSON;
import static org.deegree.services.oaf.RequestFormat.XML;
import static org.deegree.services.oaf.RequestFormat.byFormatParameter;
import static org.deegree.services.oaf.feature.FeatureResponseUtil.createGmlResponseWithHeaders;

/**
 * @author <a href="mailto:goltz@lat-lon.de">Lyn Goltz </a>
 */
@Path("/datasets/{datasetId}/collections/{collectionId}/items/{featureId}")
public class Feature {

    private final DataAccess dataAccess = DataAccessFactory.getInstance();

    @GET
    @Produces({ APPLICATION_GEOJSON })
    @Operation(summary = "retrieves feature of collection {collectionId}", description = "Retrieves one single feature of the collection with the id {collectionId}")
    @Tag(name = "Data")
    public Response featureGeoJson(
                    @Context
                                    UriInfo uriInfo,
                    @PathParam("datasetId")
                                    String datasetId,
                    @PathParam("collectionId")
                                    String collectionId,
                    @PathParam("featureId")
                                    String featureId,
                    @Parameter(description = "The coordinate reference system of the response geometries.", style = ParameterStyle.FORM)
                    @QueryParam("crs")
                                    String crs,
                    @Parameter(description = "The request output format.", style = ParameterStyle.FORM)
                    @QueryParam("f")
                                    String format )
                    throws UnknownCollectionId, InternalQueryException, InvalidParameterValue, UnknownDatasetId {
        return feature( uriInfo, datasetId, collectionId, featureId, crs, format, JSON );
    }

    @GET
    @Produces({ APPLICATION_GML })
    @Operation(summary = "retrieves feature of collection {collectionId}", description = "Retrieves one single feature of the collection with the id {collectionId}")
    @Tag(name = "Data")
    public Response featureGml(
                    @Context
                                    UriInfo uriInfo,
                    @PathParam("datasetId")
                                    String datasetId,
                    @PathParam("collectionId")
                                    String collectionId,
                    @PathParam("featureId")
                                    String featureId,
                    @Parameter(description = "The coordinate reference system of the response geometries.", style = ParameterStyle.FORM)
                    @QueryParam("crs")
                                    String crs,
                    @Parameter(description = "The request output format.", style = ParameterStyle.FORM)
                    @QueryParam("f")
                                    String format )
                    throws UnknownCollectionId, InternalQueryException, InvalidParameterValue, UnknownDatasetId {
        return feature( uriInfo, datasetId, collectionId, featureId, crs, format, XML );
    }

    @GET
    @Produces({ TEXT_HTML })
    @Operation(hidden = true)
    public Response featureHtml(
                    @Context
                                    UriInfo uriInfo,
                    @PathParam("datasetId")
                                    String datasetId,
                    @PathParam("collectionId")
                                    String collectionId,
                    @PathParam("featureId")
                                    String featureId,
                    @Parameter(description = "The coordinate reference system of the response geometries.", style = ParameterStyle.FORM)
                    @QueryParam("crs")
                                    String crs,
                    @Parameter(description = "The request output format.", style = ParameterStyle.FORM)
                    @QueryParam("f")
                                    String format )
                    throws InvalidParameterValue, UnknownDatasetId, UnknownCollectionId, InternalQueryException {
        return feature( uriInfo, datasetId, collectionId, featureId, crs, format, HTML );
    }

    @GET
    @Operation(hidden = true)
    public Response featureOther(
                    @Context
                                    UriInfo uriInfo,
                    @PathParam("datasetId")
                                    String datasetId,
                    @PathParam("collectionId")
                                    String collectionId,
                    @PathParam("featureId")
                                    String featureId,
                    @Parameter(description = "The coordinate reference system of the response geometries.", style = ParameterStyle.FORM)
                    @QueryParam("crs")
                                    String crs,
                    @Parameter(description = "The request output format.", style = ParameterStyle.FORM)
                    @QueryParam("f")
                                    String format )
                    throws InvalidParameterValue, UnknownDatasetId, UnknownCollectionId, InternalQueryException {
        return feature( uriInfo, datasetId, collectionId, featureId, crs, format, HTML );
    }

    private Response feature( UriInfo uriInfo, String datasetId, String collectionId, String featureId, String crs,
                              String formatParamValue, RequestFormat defaultFormat )
                    throws UnknownCollectionId, InternalQueryException, InvalidParameterValue, UnknownDatasetId {
        RequestFormat requestFormat = byFormatParameter( formatParamValue, defaultFormat );
        if ( HTML.equals( requestFormat ) ) {
            return Response.ok( getClass().getResourceAsStream( "/feature.html" ), TEXT_HTML ).build();
        }

        LinkBuilder linkBuilder = new LinkBuilder( uriInfo );
        FeatureResponse featureResponse = dataAccess.retrieveFeature( datasetId, collectionId, featureId, crs,
                                                                      linkBuilder );
        if ( XML.equals( requestFormat ) ) {
            return createGmlResponseWithHeaders( featureResponse );
        }
        return Response.ok( featureResponse, APPLICATION_GEOJSON ).build();
    }

}